Makefile: Add support for Windows user
Maybe someone waana test my project XD

+++ Makefile
@@ -21,9 +21,19 @@ $(OBJ_DIR)/%.o: $(SRC_DIR)/%.cpp | $(OBJ_DIR)
 	$(CXX) $(CXXFLAGS) -c $< -o $@
 
 $(OBJ_DIR):
+ifeq ($(OS),Windows_NT)
+	@if not exist $(OBJ_DIR) mkdir $(OBJ_DIR)
+else
 	mkdir -p $(OBJ_DIR)
+endif
 
 clean:
+ifeq ($(OS),Windows_NT)
+	if exist $(OBJ_DIR) rmdir /s /q $(OBJ_DIR)
+	if exist $(TARGET).exe del /q $(TARGET).exe
+	if exist $(TARGET) del /q $(TARGET)
+else
 	rm -rf $(OBJ_DIR) $(TARGET)
+endif
 
 .PHONY: all clean
